+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 1.5.1
 ## Added
 - Support for multiple slider titles on the same page via additional `header` property inside `requestOptions` entries
+- Allowed to configure `requestOptions` inside the widget settings, so there is no need anymore to configure the same settings inside the CMS page configuration and the extension config
 
 ## 1.5.0
 ## Added